rename fjroute file
